Resolve promises in parallel instead of sequentially.
Fetching information about the relevant class name for autocompletion does not depend on fetching the current class name's information. Calculating the suggestions does, but we can fire both asynchronous tasks in parallel, we only need to ensure that both are finished by the time we want to calculate the suggestions.

Determine the current class name and fetch its information